Move GroupsParser and GroupDescriptions
